pyproject.toml
[project]
name = "zyrastory-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "pillow (>=11.3.0,<12.0.0)",
    "line-bot-sdk (>=3.19.1,<4.0.0)",
    "supabase (>=2.22.0,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "redis (>=7.0.0,<8.0.0)",
    "argon2-cffi (>=25.1.0,<26.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
